Improve pushing piece placement move heuristic

+++ cpp/src/strategies/minmax/include/minmax/minmax_aliases.hpp
@@ -4,5 +4,5 @@
 
 namespace Alphalcazar::Strategy::MinMax {
 	using Depth = std::uint8_t;
-	using Score = std::int16_t;
+	using Score = std::int32_t;
 }
